LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-justify-baseline-002-expected.txt

line1
line2
line1
line2
line1
line2
line1
line2
line1
line2
line1
line2

FAIL #target > div 1 assert_equals:
<div data-offset-x="120">line1<br>line2</div>
offsetLeft expected 120 but got 0
PASS #target > div 2
PASS #target > div 3
PASS #target > div 4
PASS #target > div 5
PASS #target > div 6
